Inventory-Risk-Explainability-main/Development-Files/APIs/ChatHistory/requirements.txt
python-dotenv
fastapi
uvicorn
httpx==0.27.2
sse_starlette
urllib3
hdbcli
cfenv
sap-xssec
requests
